+++ src/main/java/com/coolGroup/org/config/BootstrapperUtility.java
@@ -3,7 +3,7 @@ package com.coolGroup.org.config;
 import com.coolGroup.org.models.Student;
 import com.coolGroup.org.models.Module;
 import com.coolGroup.org.models.dtos.PaymentAccountDto;
-import com.coolGroup.org.services.IWorker;
+import com.coolGroup.org.services.abstracts.IWorker;
 
 import java.util.ArrayList;
 import java.util.List;
