+++ src/main/java/com/coolGroup/org/controllers/api/AnalyticsController.java
@@ -1,6 +1,6 @@
-package com.coolGroup.org.controllers;
+package com.coolGroup.org.controllers.api;
 
-import com.coolGroup.org.services.IWorker;
+import com.coolGroup.org.services.abstracts.IWorker;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
