+++ src/main/java/com/coolGroup/org/controllers/api/ModuleController.java
@@ -1,10 +1,10 @@
-package com.coolGroup.org.controllers;
+package com.coolGroup.org.controllers.api;
 
 import com.coolGroup.org.config.PermissionUtility;
 import com.coolGroup.org.models.Staff;
 import com.coolGroup.org.models.Student;
 import com.coolGroup.org.models.enums.Permission;
-import com.coolGroup.org.services.IWorker;
+import com.coolGroup.org.services.abstracts.IWorker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
