+++ src/main/java/com/coolGroup/org/models/Staff.java
@@ -1,9 +1,11 @@
 package com.coolGroup.org.models;
 
+import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+@Entity
 public class Staff {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
