+++ src/main/java/com/coolGroup/org/services/abstracts/IModuleService.java
@@ -1,4 +1,4 @@
-package com.coolGroup.org.services;
+package com.coolGroup.org.services.abstracts;
 
 import com.coolGroup.org.models.Module;
 import org.springframework.stereotype.Service;
