+++ src/main/java/com/coolGroup/org/services/abstracts/IStudentService.java
@@ -1,4 +1,4 @@
-package com.coolGroup.org.services;
+package com.coolGroup.org.services.abstracts;
 
 import com.coolGroup.org.models.PaymentAccount;
 import com.coolGroup.org.models.Student;
