+++ src/main/java/com/coolGroup/org/services/concretes/StudentService.java
@@ -1,16 +1,16 @@
-package com.coolGroup.org.services;
+package com.coolGroup.org.services.concretes;
 
 import com.coolGroup.org.models.PaymentAccount;
 import com.coolGroup.org.models.Student;
 import com.coolGroup.org.models.dtos.PaymentAccountDto;
 import com.coolGroup.org.repositories.StudentRepository;
+import com.coolGroup.org.services.abstracts.IStudentService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 public class StudentService implements IStudentService {
